backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# AI & LLM Logic
ollama==0.1.6
langchain==0.1.12
langchain-ollama==0.0.1

# Utilities
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.27.0

# Database
mysql-connector-python==8.3.0

# Zero-Shot Image Classification (SigLIP)
transformers>=4.36.0
torch>=2.0.0
Pillow>=10.0.0

dotenv